Patch to allow summarized output for nagios
As nagios mostly uses NRPE a very large output is often impossible. This
patch adds "-N" option to get a summary instead of output a line for
every certificate. "-N" implies "-n" and enables Nagios mode on use.

It will output one of these variations (order expired, expire, valid):
1 certificate(s) expired (FILE:domain.tld.pem on Nov 19 2015)|days=-336
7 certificate(s) will expire (FILE:domain.tld.pem on Dec 22 2016)|days=63
11 valid certificate(s)|days=87

It also outputs the certificate which will expire first or was expired
longest time ago. Performance data is set to the minimum of days.