Update extraction of certificate issuer
The issuer can be extracted from the temporary certificate file. When using ssl-cert-check in Debian Stretch, the issuer will not be extracted, because the regular expression does not fit anymore when using openssl command. The format of the temporary certificate file is the same as in previous Debian releases, therefore the regular expression to extract the issuer still fits and grep does the job.